UPSTREAM: pstore: Use global ftrace filters for function trace filtering
Currently, pstore doesn't have any filters setup for function tracing.
This has the associated overhead and may not be useful for users looking
for tracing specific set of functions.

ftrace's regular function trace filtering is done writing to
tracing/set_ftrace_filter however this is not available if not requested.
In order to be able to use this feature, the support to request global
filtering introduced earlier in the series should be requested before
registering the ftrace ops. Here we do the same.

BUG=chromium:633355
TEST=Boot 4.4 kernel on veyron or kevin/gru

Change-Id: Ic000a7d4a4f598777cdfa707cacfe84ccbc8c653
Signed-off-by: Joel Fernandes <joelaf@google.com>
Signed-off-by: Kees Cook <keescook@chromium.org>
Signed-off-by: Guenter Roeck <groeck@chromium.org>
(cherry picked from commit 7a0032f50472)
Reviewed-on: https://chromium-review.googlesource.com/440558
Signed-off-by: Hans-Gerhard Weinreich <hans@ip-172-31-43-104.eu-west-1.compute.internal>