CHROMIUM: drm/rockchip: psr: Sanitize semantics of allow/disallow API
Currently both rockchip_drm_psr_activate() and _deactivate() only set the
boolean "active" flag without actually making sure that hardware state
complies with it.

Since we are going to extend the usage of this API to properly lock PSR
for the duration of atomic commits, we change the semantics in following
way:
 - a counter is used to track the number of disallow requests,
 - PSR is actually disabled in hardware on first disallow request,
 - PSR enable work is scheduled on last disallow request.

The above allows using the API as a way to deterministically synchronize
PSR state changes with other DRM events, i.e. atomic commits and cursor
updates. As a nice side effect, the naming is sorted out and we have
"inhibit" for stopping the software logic and "enable" for hardware
state.

BUG=chrome-os-partner:62479
TEST=Pick CL:436410, no more iommu errors upon screen wakeup

Change-Id: Ibc9c139806e294e32e9086fc3d4d39c2df4f9ab4
Signed-off-by: Tomasz Figa <tfiga@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/438228
Tested-by: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Kristian H. Kristensen <hoegsberg@chromium.org>
Signed-off-by: Hans-Gerhard Weinreich <hans@ip-172-31-43-104.eu-west-1.compute.internal>