CHROMIUM: config: Enable wdt87xx_i2c driver
This driver is used for Weida touchscreens like that used in Sumo.  As
Sumo has been moved to kernel 4.4, this driver must be compiled into the
4.4 kernel or the touchscreen will neither work nor update.  This CL
just adds the wdt87xx_i2c driver.

BUG=chromium:690661
TEST=I rebuilt the kernel and deployed on my Sumo.  With the new kernel,
the driver was listed under /sys/bus/i2c/drivers, the touchscreen
worked, and the touch firmware updater saw the screen and the logs
were as expected.

Change-Id: I4b6f36367a08de9c0d7108779d385f0750e16034
Signed-off-by: Charlie Mooney <charliemooney@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/442095
Reviewed-by: Benson Leung <bleung@chromium.org>
Signed-off-by: Hans-Gerhard Weinreich <hans@ip-172-31-43-104.eu-west-1.compute.internal>